ratings are changed to integers because 5 stars

+++ database/migrations/2022_02_20_183525_create_ratings_table.php
@@ -15,7 +15,7 @@ class CreateRatingsTable extends Migration
     {
         Schema::create('ratings', function (Blueprint $table) {
             $table->foreignId('reservation_id')->constrained()->onDelete('cascade')->onUpdate('cascade');
-            $table->decimal('rating');
+            $table->integer('rating');
 
             $table->timestamps();
         });
